add notification senders for Discord, Email, and Slack with separation; refactor uptime type to Double

+++ shared/src/commonMain/kotlin/pt/isel/api_pm/dto/metric/AggregatedMetric.kt
@@ -13,7 +13,7 @@ data class AggregatedMetric (
     val totalRequests: Long,
     val errorRate: Double,
     val throughput: Long,
-    val uptime: Long,
+    val uptime: Double,
     val percentile95: Long,
     val percentile99: Long,
     val statusCodeDistribution: Map<Int, Long>
